virtcontainers: kata_agent: Move out a generic function
The specific agent implementation kata_agent was defining a very
useful generic function that is now moved to the global file
network.go.

Fixes #629
