+++ virtcontainers/network.go
@@ -27,6 +27,7 @@ import (
 	"github.com/vishvananda/netns"
 	"golang.org/x/sys/unix"
 
+	"github.com/kata-containers/agent/protocols/grpc"
 	"github.com/kata-containers/runtime/virtcontainers/device/config"
 	"github.com/kata-containers/runtime/virtcontainers/device/drivers"
 	"github.com/kata-containers/runtime/virtcontainers/pkg/uuid"
@@ -1338,6 +1339,93 @@ func createVirtualNetworkEndpoint(idx int, ifName string, interworkingModel NetI
 	}
 
 	return endpoint, nil
+}
+
+func generateInterfacesAndRoutes(networkNS NetworkNamespace) ([]*grpc.Interface, []*grpc.Route, error) {
+
+	if networkNS.NetNsPath == "" {
+		return nil, nil, nil
+	}
+
+	var routes []*grpc.Route
+	var ifaces []*grpc.Interface
+
+	for _, endpoint := range networkNS.Endpoints {
+
+		var ipAddresses []*grpc.IPAddress
+		for _, addr := range endpoint.Properties().Addrs {
+			// Skip IPv6 because not supported
+			if addr.IP.To4() == nil {
+				// Skip IPv6 because not supported
+				networkLogger().WithFields(logrus.Fields{
+					"unsupported-address-type": "ipv6",
+					"address":                  addr,
+				}).Warn("unsupported address")
+				continue
+			}
+			// Skip localhost interface
+			if addr.IP.IsLoopback() {
+				continue
+			}
+			netMask, _ := addr.Mask.Size()
+			ipAddress := grpc.IPAddress{
+				Family:  grpc.IPFamily_v4,
+				Address: addr.IP.String(),
+				Mask:    fmt.Sprintf("%d", netMask),
+			}
+			ipAddresses = append(ipAddresses, &ipAddress)
+		}
+		ifc := grpc.Interface{
+			IPAddresses: ipAddresses,
+			Device:      endpoint.Name(),
+			Name:        endpoint.Name(),
+			Mtu:         uint64(endpoint.Properties().Iface.MTU),
+			HwAddr:      endpoint.HardwareAddr(),
+		}
+
+		ifaces = append(ifaces, &ifc)
+
+		for _, route := range endpoint.Properties().Routes {
+			var r grpc.Route
+
+			if route.Dst != nil {
+				r.Dest = route.Dst.String()
+
+				if route.Dst.IP.To4() == nil {
+					// Skip IPv6 because not supported
+					networkLogger().WithFields(logrus.Fields{
+						"unsupported-route-type": "ipv6",
+						"destination":            r.Dest,
+					}).Warn("unsupported route")
+					continue
+				}
+			}
+
+			if route.Gw != nil {
+				gateway := route.Gw.String()
+
+				if route.Gw.To4() == nil {
+					// Skip IPv6 because is is not supported
+					networkLogger().WithFields(logrus.Fields{
+						"unsupported-route-type": "ipv6",
+						"gateway":                gateway,
+					}).Warn("unsupported route")
+					continue
+				}
+				r.Gateway = gateway
+			}
+
+			if route.Src != nil {
+				r.Source = route.Src.String()
+			}
+
+			r.Device = endpoint.Name()
+			r.Scope = uint32(route.Scope)
+			routes = append(routes, &r)
+
+		}
+	}
+	return ifaces, routes, nil
 }
 
 func networkInfoFromLink(handle *netlink.Handle, link netlink.Link) (NetworkInfo, error) {
